requirements.txt
streamlit
pandas
scikit-learn
joblib
mysql-connector-python